Updates to allow for CORS security
Since this project was developed, CORS security has become standard in web browsers. Direct browser to server communication is blocked, and the ways to solve this are either to disable browser security, which could post a danger, or to use a server and server to server communication instead. As github pages do not support web server hosting, I have made use of a free and open source third party proxy server allOrigins.win.

There is also a CORS proxy server written in python that could be used to run this project locally or hosted elsewhere.